requirements.txt
commonroad-io==2021.1
sumocr==2021.1
jupyter
matplotlib
numpy
pyyaml
